feat!: hardcode oauth client ids
Removes support for the build-time `BUILTIN_CLIENT_IDS` environment
variable and replaces it with a simple match statement mapping hostnames
to client IDs.

The point of that var was so that anyone packaging fj for a package
manager would be able to include IDs for any associated instances, i.e.
a linux distro that uses Forgejo for its forge.

However this was overkill. Someone packaging it can just patch that file
to add their own IDs, and doing it this way prevented oauth from being
supported on from-source builds (such as `cargo build`).

My worry that the client IDs being publicly visible was a risk was also
unfounded, and not an issue.