feat: `--autofill` flag on `pr create`
A new `--autofill` flag for `pr create` that automatically populates the
title and body from the included commits. If there's only one commit,
the PR matches the title and body. If there's multiple the title is the
branch name, and the body is a definition list containing each's
commit's message.

As a test that it works, this very PR was written using this new
feature!

Implements #217