feat(issue): read body from file
The current command to create issues accepts the body either on the
command line or by opening a text editor.

I like to write bodies in local Markdown files, so that I can write a
few offline before uploading them. Even if I am online, having the issue
saved on disk gives me more confidence that it won't be lost should
forgejo-cli encounter an error.

This adds the `--body-file` flag to read the body from a file. This flag
can’t be combined with the `--body` argument. This is enforced at
runtime by Clap. If both are specified, it will fail with:

```
error: the argument '--body <BODY>' cannot be used with '--body-file <BODY_FILE>'
```

If the file is missing, the error looks like this:
```
Error: Error reading file `/tmp/f.md`

Caused by:
    No such file or directory (os error 2)

Location:
    src/issues.rs:263:26
```